refactor(api): remove unused formatUserResponse and formatAgentResponse functions
