+++ web-app/src/lib/api/formatters/index.ts
@@ -1,0 +1,2 @@
+export * from "./agent";
+export * from "./user";
