+++ web-app/src/lib/api/formatters/user.ts
@@ -1,0 +1,26 @@
+import "server-only";
+
+import { UserResponse, userResponseWrapperSchema } from "@/lib/api/schemas";
+import { dateToISO } from "@/lib/api/utils";
+import { User } from "@/prisma/generated/client";
+
+/**
+ * Formats user data for API response
+ */
+export function formatUserResponse(user: User): UserResponse {
+  const formatted = {
+    user: {
+      id: user.id,
+      createdAt: dateToISO(user.createdAt),
+      updatedAt: dateToISO(user.updatedAt),
+      name: user.name,
+      email: user.email,
+      termsAccepted: user.termsAccepted,
+      marketingOptIn: user.marketingOptIn,
+      stripeCustomerId: user.stripeCustomerId,
+    },
+  };
+
+  // Validate the formatted response
+  return userResponseWrapperSchema.parse(formatted);
+}
